"Classe Cliente Modelo - Atualização"

+++ API/ParqueAPI/Models/Cliente.cs
@@ -10,9 +10,9 @@ namespace ParqueAPI.Models
         public long ClienteID { get; set; }
         public string NomeCliente { get; set; }
         public string EmailCliente { get; set; }
-        public int NifCliente { get; set; }
+        public long NifCliente { get; set; }
         public string MetodoPagamento { get; set; }
-        public int Credito { get; set; }
+        public float Credito { get; set; }
 
     }
 }
